Created Modules for image and quotes
